+++ Schedule_Calendar/page_style.css
@@ -1,0 +1,146 @@
+#cover {
+    width:1200px;
+    background-color: #afeeee;
+    background-repeat:repeat-y;
+    margin-right:auto;
+    margin-left:auto;
+}
+
+#contents {
+    width: 1000px;
+    margin-right:auto;
+    margin-left:auto;
+    background-color: #f7a5a5;
+}
+
+.square_btn_a {
+    width: 130px;
+    height: 25px;
+    display: inline-block;
+    padding: 5px 5px;
+    text-decoration: none;
+    background: #f9f7e5;
+    color: #000000;
+    border-bottom: solid 4px #ba55d3;
+    border-radius: 10px;
+}
+
+.square_btn_a:active {
+    -ms-transform: translateY(4px);
+    -webkit-transform: translateY(4px);
+    transform: translateY(4px);
+    box-shadow: 0px 0px 1px rgba(0, 0, 0, 0.2);
+    border-bottom: none;
+}
+
+.square_btn_b {
+    width: 130px;
+    height: 25px;
+    display: inline-block;
+    padding: 5px 5px;
+    text-decoration: none;
+    background: #f9f7e5;
+    color: #000000;
+    border-bottom: solid 4px #20b2aa;
+    border-radius: 10px;
+}
+
+.square_btn_b:active {
+    -ms-transform: translateY(4px);
+    -webkit-transform: translateY(4px);
+    transform: translateY(4px);
+    box-shadow: 0px 0px 1px rgba(0, 0, 0, 0.2);
+    border-bottom: none;
+}
+
+.square_btn_c {
+    width: 130px;
+    height: 25px;
+    display: inline-block;
+    padding: 5px 5px;
+    text-decoration: none;
+    background: #f9f7e5;
+    color: #000000;
+    border-bottom: solid 4px #ffff00;
+    border-radius: 10px;
+}
+
+.square_btn_c:active {
+    -ms-transform: translateY(4px);
+    -webkit-transform: translateY(4px);
+    transform: translateY(4px);
+    box-shadow: 0px 0px 1px rgba(0, 0, 0, 0.2);
+    border-bottom: none;
+}
+
+h1 {
+    color: red;
+    font-family: "arial black";
+}
+
+.title-shadow { 
+    text-shadow: 3px 3px 5px #0000ff; 
+}
+
+table.calen { 
+    border-collapse:collapse; 
+    font-size:11pt; 
+    background-color:#f7d2d5;
+}
+table.calen th { 
+    border:1px solid #1115e0;
+    height:30px; 
+}
+table.calen td { 
+    border:1px solid #1115e0;
+    text-align:center; 
+    height:30px; 
+    width:30px; 
+}
+table.calen .sat { 
+    color:blue; 
+}
+table.calen .sun { 
+    color:red; 
+}
+table.calen .today { 
+    background-color:#ff7f16;
+    font-weight:bold; 
+}
+table.calen tr.bg1 { 
+    background-color:#ffc493;
+}
+table.calen tr.bg2 { 
+    background-color:#fff; 
+}
+
+.square_btn {
+    width: 130px;
+    height: 25px;
+    display: inline-block;
+    padding: 5px 5px;
+    text-decoration: none;
+    background: #f9f7e5;
+    color: #000000;
+    border-bottom: solid 4px #ed0000;
+    border-radius: 10px;
+}
+
+.square_btn:active {
+    -ms-transform: translateY(4px);
+    -webkit-transform: translateY(4px);
+    transform: translateY(4px);
+    box-shadow: 0px 0px 1px rgba(0, 0, 0, 0.2);
+    border-bottom: none;
+}
+
+#footer {
+    clear:both;
+    text-align:center;
+    font-size:13px;
+    color: #999999;
+    margin-top:20px;
+    margin-right:10px ;
+    margin-left:10px;
+    border-top:1px dotted #ed0000;
+}
